academic-research-helper/requirements.txt
qdrant-client
python-dotenv
pypdf
orjson
uvloop; platform_system != 'Windows'
euriai
rich

